mark 3, by gustavhus

+++ src/route_cal.h
@@ -7,8 +7,8 @@
 
 extern int array_size;
 int route_cal(int array[array_size][array_size], int start, int end);
-double plane_cal(int departure, int arrival);
-void travel_comparison(double plane, int train);
+int plane_cal(int departure, int arrival);
+void travel_comparison(int plane, int train);
 void change_route_time(int array[array_size][array_size]);
 void print_line();
